+++ py-rattler/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Expose `write_repodata` in py-rattler for repodata generation from record metadata
+
 ## [0.23.2] - 2026-03-19
 
 ### Added